Check for extra-long run lengths
I'm not sure about this -- the wikipedia page says that I should be
able to enforce stricter constraints on this, but when I actually
attempt to impose those constraints the tests fail.  I'm probably just
reading the spec wrong, but LVDS looks like a strictly better spec
anyway so I'm just going to try and implement that.