Add an 8b/10b decoder
There's a test for this, but it's not super greeat -- all it actually
does is check that data words are correctly propogated, it doesn't
look at control words.  I figure it's best to leave the control words
to the controller, as testing how those work is going to require some
integration with the controller.